+++ webpack/dev_h5.config.js
@@ -4,16 +4,19 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 normalConfig.plugins.push(
-    new HtmlWebpackPlugin({
-        template: '/public/template/index.html',
-        favicon: path.resolve('./public/favicon.ico'),
-    })
+  new HtmlWebpackPlugin({
+    template: path.resolve("./public/template/index-h5.html"),
+    favicon: path.resolve("./public/favicon.ico"),
+  })
 );
 
-module.exports = merge(normalConfig, {
-    entry: path.resolve(__dirname, '../src/pages/PC/src/index.tsx'), // string | object | array
+const config = merge(normalConfig, {
+    entry: path.resolve(__dirname, '../src/pages/H5/src/index.tsx'), // string | object | array
     output: {
         path: path.resolve(__dirname, '../dist'),
         filename: '[name].[contenthash].js',
     },
 });
+
+
+module.exports = config;
